Added some functionality for routes and geolocation dot
Routes now appear, with correct colors and widths, on the screen after moving the map little bit. This is still a WIP. The states must be modified so the map can loaded immediately. There should also be some logic in case the map can't be loaded or the data can't be pulled from 'https://shuttles.rpi.edu/routes'.